@@ -14,12 +14,13 @@
   "prefer-stable": true,
   "minimum-stability": "dev",
   "require": {
-    "php": ">=7.4",
+    "php": ">=8.1",
     "ext-json": "*",
-    "symfony/yaml": "> 4"
+    "symfony/yaml": "> 4",
+    "ext-simplexml": "*"
   },
   "require-dev": {
-    "phpunit/phpunit": "5.7.*|7.4.*|^9.5"
+    "phpunit/phpunit": "^9.6"
   },
   "suggest": {
     "ext-curl": "*"
