@@ -7,15 +7,15 @@ use ByJG\Serializer\SerializerObject;
 class PlainTextFormatter implements FormatterInterface
 {
 
-    protected $breakLine = "\n";
-    protected $startOfLine = "";
-    protected $ignorePropertyName = true;
+    protected string $breakLine = "\n";
+    protected string $startOfLine = "";
+    protected bool $ignorePropertyName = true;
 
     /**
-     * @param array $serializable
+     * @param array|object $serializable
      * @return mixed
      */
-    public function process($serializable)
+    public function process(array|object $serializable): string
     {
         if (!is_array($serializable)) {
             return $this->processInternal(SerializerObject::instance($serializable)->serialize());
@@ -25,7 +25,7 @@ class PlainTextFormatter implements FormatterInterface
         
     }
 
-    protected function processInternal($serializable)
+    protected function processInternal(array|object $serializable): string
     {
         $result = "";
 
