@@ -21,7 +21,7 @@ abstract class BaseModel extends BindableObject
         }
     }
 
-    public function toArray()
+    public function toArray(): array
     {
         return SerializerObject::instance($this)->serialize();
     }
