@@ -7,7 +7,7 @@ use ByJG\Serializer\PropertyPattern\PropertyPatternInterface;
 
 interface BindableInterface
 {
-    public function bindFrom($source, ?PropertyPatternInterface $propertyPattern = null);
+    public function bindFrom(array|object $source, ?PropertyPatternInterface $propertyPattern = null): void;
 
-    public function bindTo($target, ?PropertyPatternInterface $propertyPattern = null);
+    public function bindTo(array|object $target, ?PropertyPatternInterface $propertyPattern = null): void;
 }
