@@ -9,10 +9,10 @@ class YamlFormatter implements FormatterInterface
 {
 
     /**
-     * @param array|object $serializable
+     * @param object|array $serializable
      * @return string
      */
-    public function process($serializable)
+    public function process(object|array $serializable): string
     {
         if (is_array($serializable)) {
             return Yaml::dump($serializable, 2, 2);
